# Add perf tests for mapping buffers to NOC