… tests/api/test_cluster.cpp