### Issue

#1272 

### Description

Add tests for profiling pinning/unpinning and mapping buffers to NOC by using IOCTL calls to KMD. There are 3 new tests. 

- MapDifferentSizesBufferToNOC - point of the test is the similar to one that does mapping of different sizes, just this one is doing mapping to NOC as well, not just for IOVA
- Map1GBPages - map 3 pages where each one is 1GB in size. This should simulate metal use case with buffers (not mapping 4 because that won't fit because of 32 bit limit and there are already some small buffers mapped)
- Map1GBPagesSysmemBuffers - same as the previous one just to confirm there is no overhead of going through SysmemBuffer class

The most important thing here is mapping of 1GB pages. Results are quite similar if we go through SysmemBuffer class or not

```bash
Average map time for 1GB page: 1.24075e+07 ns. (0.0124075 seconds)
Average unmap time for 1GB page: 5.14929e+06 ns. (0.00514929 seconds)
```

So not really sure where the overhead metal infra guys have seen is coming from

### List of the changes

- Add 3 tests to IOMMU benchmark suite
- Small cleanup of unused varibales

### Testing
Local testing with KMD 2.0 and IOMMU enabled

### API Changes
/